+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/base/ButtonTag.java
@@ -12,7 +12,8 @@
  * Lesser General Public License for more details.
  * 
  * You should have received a copy of the GNU Lesser General Public
- * License along with this library. */
+ * License along with this library.
+ */
 package com.jpexs.decompiler.flash.tags.base;
 
 import com.jpexs.decompiler.flash.ReadOnlyTagList;
@@ -78,6 +79,11 @@ public abstract class ButtonTag extends DrawableTag implements Timelined {
     @Override
     public RECT getRect() {
         return getRect(new HashSet<>());
+    }
+
+    @Override
+    public RECT getRectWithStrokes() {
+        return getRect();
     }
 
     @Override