+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 - AS3 - Remove trait which is outside class
 
 ### Fixed
-- Flash viewer - bitmap line style
+- Flash viewer - bitmap stroke style, strokes scaling, cropped strokes
 
 ## [14.2.1] - 2021-03-13
 ### Added